Calibrate motor cannon on helper instantiation
This applies a low power to the cannon motor on initialization, and
waits until the motor stops moving. This power is low enough so a bullet
is not fired, which means it will stop just before firing the bullet.
Afterwards we move the motor a bit backwards, so the bullet is loosened.

The point of doing this is to set the cannon in a determined position.
That way, we know that rotating the gearing one round (three turns of
the motor) will fire exactly one bullet.