Call listeners when the sensor registers a new color
This stores the last color registered in RobotState. When the color
received from the sensors differs from the last color, the state
listeners are triggered.

Currently, only the avoidEdge method is triggered when the new color is
black. We will expand this to include all the colors we use and all the
methods of the StateReceiver.